% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_rho.R
\name{sibcorr_bs}
\alias{sibcorr_bs}
\title{Estimate sibling and cousin correlations}
\usage{
sibcorr_bs(data, id1, id2, id3, ..., cousins = FALSE, reps = 50,
  ci_level = 95)
}
\arguments{
\item{data}{Estimation data set.}

\item{id1}{Identifier for the children}

\item{id2}{Identifier for the parents}

\item{id3}{Identifier for the grandparents}

\item{...}{Other options for \code{\link{sibcorr}}}

\item{cousins}{Estimate cousin correlation if TRUE, otherwise sibling correlation.}

\item{reps}{Number of bootstrap replications.}

\item{ci_level}{Set level for bootstrap confidence interval.}
}
\description{
Estimate sibling and cousin correlations
}
