% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibcorr.R
\name{sibcorr}
\alias{sibcorr}
\title{Estimate sibling and cousin correlations.}
\usage{
sibcorr(formula, data, weight = 4)
}
\arguments{
\item{formula}{Three-part formula describing the outcome variable, control variables to be regressed out, and identifiers for the individual, immediate family, and extended family}

\item{data}{Estimation data set.}

\item{weight}{Select one of four weighting schemes.}
}
\value{
The estimated sibling or cousin correlation coefficient
and number of individuals, sibling or cousin pairs, and families or extended families
}
\description{
Estimate sibling and cousin correlations.
}
