% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibcorr.R
\name{sibcorr_bs}
\alias{sibcorr_bs}
\title{Estimate sibling and cousin correlation}
\usage{
sibcorr_bs(data, weight = 4, controls = NULL, cousins = FALSE,
  reps = 50, ci_level = 95)
}
\arguments{
\item{data}{Estimation data set.}

\item{weight}{Select one of four weighting schemes.}

\item{controls}{Control variables to regress out before estimating correlation.}

\item{cousins}{Estimate cousin correlation if TRUE, otherwise sibling correlation.}

\item{reps}{Number of bootstrap replications.}

\item{ci_level}{Set level for bootstrap confidence interval.}
}
\description{
Estimate sibling and cousin correlation
}
