% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibcorr_bs.R
\name{sibcorr_bs}
\alias{sibcorr_bs}
\title{Estimate sibling and cousin correlations with boostrap confidence intervals.}
\usage{
sibcorr_bs(formula, data, ..., reps = 50, ci_level = 95, cores = 1)
}
\arguments{
\item{formula}{Three-part formula describing the outcome variable, control variables to be regressed out, and identifiers for the individual, immediate family, and extended family}

\item{data}{Estimation data set.}

\item{...}{Other options for \code{\link{sibcorr}}}

\item{reps}{Number of bootstrap replications}

\item{ci_level}{Set level for bootstrap confidence interval}

\item{cores}{Set number of processor cores to use for bootstrap estimation}
}
\value{
The estimated sibling or cousin correlation coefficient,
number of individuals, sibling or cousin pairs, and families or extended families,
and bootstrap confidence interval
}
\description{
\code{sibcorr_bs} estimates sibling or cousin correlations with block bootstrap standard errors.
}
