% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibcorr_bs.R
\name{sibcorr_bs}
\alias{sibcorr_bs}
\title{Estimate sibling and cousin correlations}
\usage{
sibcorr_bs(formula, data, ..., cousins = FALSE, reps = 50, ci_level = 95,
  cores = 1)
}
\arguments{
\item{formula}{Three-part formula describing the outcome variable, control variables to be regressed out, and identifiers for the individual, immediate family, and extended family}

\item{data}{Estimation data set.}

\item{...}{Other options for \code{\link{sibcorr}}}

\item{cousins}{Estimate cousin correlation if TRUE, otherwise sibling correlation.}

\item{reps}{Number of bootstrap replications}

\item{ci_level}{Set level for bootstrap confidence interval}

\item{cores}{Set number of processor cores to use for bootstrap estimation}
}
\description{
Estimate sibling and cousin correlations
}
