% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibcorr_bs.R
\name{sibcorr_bs}
\alias{sibcorr_bs}
\title{Estimate sibling and cousin correlations}
\usage{
sibcorr_bs(data, id1, id2, id3 = "", ..., cousins = FALSE, reps = 50,
  ci_level = 95, cores = 1)
}
\arguments{
\item{data}{Estimation data set.}

\item{id1}{Individual identifier}

\item{id2}{Family identifier}

\item{id3}{Extended family (cousin group) identifier}

\item{...}{Other options for \code{\link{sibcorr}}}

\item{cousins}{Estimate cousin correlation if TRUE, otherwise sibling correlation.}

\item{reps}{Number of bootstrap replications}

\item{ci_level}{Set level for bootstrap confidence interval}

\item{cores}{Set number of processor cores}
}
\description{
Estimate sibling and cousin correlations
}
